test: merged demo-properties into new component based test structure
