+++ example/demo.py
@@ -11,12 +11,13 @@ import time
 from visdom import Visdom
 import argparse
 from components.text import text_basic, text_update, text_callbacks, text_close
-from components.image import image_basic, image_callback, image_save_jpeg, image_history, image_grid, image_svg
+from components.image import image_basic, image_callback, image_callback2, image_save_jpeg, image_history, image_grid, image_svg
 from components.plot_scatter import plot_scatter_basic, plot_scatter_update_opts, plot_scatter_append, plot_scatter_3d, plot_scatter_custom_marker, plot_scatter_custom_colors, plot_scatter_add_trace, plot_scatter_text_labels_1d, plot_scatter_text_labels_2d
 from components.plot_bar import plot_bar_basic, plot_bar_stacked, plot_bar_nonstacked, plot_bar_histogram, plot_bar_piechart
 from components.plot_surface import plot_surface_basic, plot_surface_basic_withnames, plot_surface_append, plot_surface_append_withnames, plot_surface_remove, plot_surface_remove_withnames, plot_surface_replace, plot_surface_replace_withnames, plot_surface_contour, plot_surface_3d
 from components.plot_line import plot_line_basic, plot_line_multiple, plot_line_webgl, plot_line_update_webgl, plot_line_update, plot_line_opts, plot_line_opts_update, plot_line_stackedarea, plot_line_maxsize, plot_line_doubleyaxis, plot_line_pytorch, plot_line_stem
 from components.plot_special import plot_special_boxplot, plot_special_quiver, plot_special_mesh, plot_special_graph
+from components.properties import properties_basic, properties_callbacks
 from components.misc import misc_plot_matplot, misc_plot_latex, misc_plot_latex_update, misc_video_tensor, misc_video_download, misc_audio_basic, misc_audio_download, misc_arbitrary_visdom, misc_getset_state
